Update patch set 4
Patch Set 4: -Code-Review

The 42-openflowplugin-new.xml has been deprecated. We have the openflowplugin.xml in its place. So we need to modify some xml that can be configured at runtime. But i donot see any xml being modified as a part of this change, so could you please explain as to how can one turn it on by demand

Patch-set: 4