Update patch set 1
Patch Set 1:

I too agree with you Jozef, but for situations like the port-status message arriving before the master is elected for the switch, we need to retry.

I can understand why DeviceContext might not be correct place, but then we need to retry them and throw up in case the error persists. Can we then put the logic in txnchainmgr instead ?

Patch-set: 1