Update patch set 2
Patch Set 2:

It fails because of the missing artifact which is created by the openflowjava patch. The job will never success as you revert the patch that create the artifact. Not sure how we can handle this, May you can remerge the openflowjava patch  and wait couple of hours till the artifact appear at nexus.

15:10:05 Error resolving artifact org.opendaylight.openflowjava:openflowjava-blueprint-config:jar:0.9.0-SNAPSHOT: Could not find artifact org.opendaylight.openflowjava:openflowjava-blueprint-config:jar:0.9.0-SNAPSHOT

Patch-set: 2
Label: Verified=0