Update patch set 1
Patch Set 1: Code-Review-1

I am not fully convinced to make changes like this. I don't think we should solve problems for other projects offset:2. I think we should just deliver messages to offset 0 projects and back. If there is a problem to establish an cluster environment the project need to be aware of problem on DS, akka whatever, but not to solve with retries on each action. :)
At least if we need (still not convinced) retries on txChain lets do it in txChainManager not in device context ...

Patch-set: 1
Label: Code-Review=-1