Update patch set 14
Patch Set 14: Code-Review-1

(3 comments)

Hi Shuva, Looking at the problem I think it's happening because while reading the flow stats plugin is storing the key to the hashmap. So if the similar flow comes after the first flow, it will find the key in the key hash map

https://github.com/opendaylight/openflowplugin/blob/7f4b511776bcdeab7d965ec78fc4a6d5b626ef47/openflowplugin-impl/src/main/java/org/opendaylight/openflowplugin/impl/statistics/StatisticsGatheringUtils.java#L285

I think here we should just retrieve the flow-id using retriveFlowId() and if there is no flow-id, we should use the alien id to augment the flow. I think this should fix the issue. Can you please try this if it works for you. Apart from that i liked the changes of storing the key when rpc is successful.

Patch-set: 14
Label: Code-Review=-1