Update patch set 1
Patch Set 1: Verified-1

Build Failed 

https://jenkins.opendaylight.org/releng/job/openflowplugin-verify-boron-mvn33-openjdk8/182/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/openflowplugin-verify-boron-mvn33-openjdk8/182/ )

https://jenkins.opendaylight.org/releng/job/openflowplugin-validate-autorelease-boron/1/ : FAILURE

https://jenkins.opendaylight.org/releng/job/openflowplugin-distribution-check-boron/5/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/openflowplugin-distribution-check-boron/5/ )

Patch-set: 1
Label: Verified=-1