Update patch set 4
Patch Set 4: Code-Review-1

(2 comments)

minor comments.

We need to run singled node CSIT test against this patch, to make sure it's not breaking existing functionality. The way DataTreeChangeListener provide the change data is different from DataChangeListener, so we need to make sure we are not breaking in this translation.

Patch-set: 4