Update patch set 3
Patch Set 3: Code-Review-1

the fix of appending the priority to the flowid to guarantee its uniqueness is not a good fix.
As we see, the flowid of both the flows are same, hence this issue occurs. The flowRegistryKey generation and comparison happens correctly, Is it possible to handle it at the application level itself?

As i see in the l2switch code, the ProactiveFloodFlowWriter and InitialFlowWriter have independent atomic counters, can we unify it to guarantee unique flow-id

Patch-set: 3
Label: Code-Review=-1