Update patch set 8
Patch Set 8:

Your patch will solve only half of the problem, because on master, single layer serialization was enabled by default, and so MeterConvertor and MeterConfigStatsResponseConvertor are not used unless flag for single layer serialization is disabled or device is using lower version of OpenFlow than 1.3. But you are right, I should make patch on top of yours then where I will add some changes to MeterMessageSerializer and MultipartReplyMeterConfigDeserializer to make them work with your changes. Because without changes in these 2 classes, as I said before, meter experimenters will not work with single layer serialization, so I think it will be still part of bug 6766.

Patch-set: 8