Update patch set 2
Patch Set 2:

> > Unfortunately we do not have much test data in nitrogen because
 > > distribution test was pulling karaf 4, however you can see in the
 > > patch test job the cluster test is far from stable:
 > > https://jenkins.opendaylight.org/releng/view/CSIT-3node/job/openflowplugin-csit-3node-gate-clustering-only-nitrogen/
 > 
 > okay, so looks like it's breaking both master and stable/carbon
 > branches

I will create rever on master then and run test-openflowplugin-core on it, so we can be sure about it. I was running test-openflowplugin-core on that patch a checking it to be sure it is not breaking anything, and according to jenkins it was not, so it is suprising for me that it is actually breaking stuff, but according to jenkins right now, that must be the case. By the way, this patch should be rebased anyway too, because of bump versions commit that was added on carbon, so we can check actual results of tests.

Patch-set: 2