Update patch set 2
Patch Set 2:

> > > > > Also, that new configuration properties in
 > openflowplugin.cfg
 > > > > will
 > > > > > throw 'Unsupported configuration property '{}={}'' warnings
 > > to
 > > > > > console because plugin will think that this properties are
 > > not
 > > > > > supported because they are not defined in OpenFlowPluginConfigurationService.PropertyType.
 > > > >
 > > > > Good catch. I think we should not throw this warning, because
 > > > > openflowplugin.cfg will be used by all the application
 > present
 > > in
 > > > > the openflowplugin repo, so plugin should just read it's
 > > > > configuration and ignore other config. I was thinking of
 > > putting
 > > > > the configuration file in the FRM, but if we follow that
 > > practice
 > > > > and individual application will have it's own config file,
 > that
 > > > > will create lots of configuration file and it's going to be
 > > pain
 > > > > for user to set different knobs in different files.
 > > >
 > > > True, but how we will warn users about incorrect configuration
 > > > then? I added that warning mainly because of high possibility
 > of
 > > > typos in such configuration files. Maybe better solution will
 > be
 > > > making configuration system extensible and extract it as
 > > blueprint
 > > > service, so for example application can add it as dependency
 > and
 > > > register it's own configuration options in it.
 > >
 > > I think we should publish clearly what configuration is been
 > picked
 > > up by plugin rather than notifying all possible wrong
 > configuration
 > > done in the config file. It should print a clear log that the
 > > "config is read and value x is set for it" That will give clear
 > > idea to user . Also I think we should avoid any config subsystem
 > > based solution :).
 > 
 > Tomas, do you have any more comments? Please read my response
 > above, in case you missed it.

Oh yes, sorry, missed the last one. I was not talking about config subsystem there, I was talking about extracting OpenFlowPluginConfigurationService via blueprint so frm can depend on it and add method there like registerKnownProperty or so that it will not produce warnings. And currently we are showing what configuration was loaded from .cfg file (you was also reviewing this change if I remmeber correctly) but also producing warning for unknown configuration properties there. Removing this warning is also solution as you suggested, but still atleast that needs to be done, because your current implementation will produce warnings in log because of that. And also, my original point with just commenting out your additions to openflowplugin.cfg should be done (like other things that are there but commented out for documentation purposes). After these 2 things will be addresed I will be fine with this patch, because rest looks okay.

Patch-set: 2