Update patch set 24
Patch Set 24:

Anil, currently this patch ended as just small cleanup, and it is not even working properly. I was thinking about abandoning it, but this patch gets rid of initialStatisticsGathering, what is currently just passing initial flag to deleteAllKnownFlows, and that method is never called, because for flow collection we are using on the fly gathering, so initialStatisticsGathering can just be statisticsGathering. But sure, I can create improvement bug on BugZilla for this.

Patch-set: 24
Label: Verified=0