Update patch set 11
Patch Set 11:

Anil, I am not using cm:property-placeholder because it will not guarantee order of setting these values. They need to be set after values from Yang file are set. I was initially trying to use property-placeholder too, but I think it was not working properly when I was testing it.

Patch-set: 11