Update patch set 2
Patch Set 2:

> tomas, so looks like it's a regression for openflow 1.0 devices? If
 > they want to use it with openflow 1.0 device, they will have to
 > disable the use-single-layer-serialization=false?

No this has nothing to do with single-layer-serialization, even when disabling it it would not fix it. OF1.0 devices are not using single-layer-serialization anyway, for them it is disabled by default. Convertors what I added was just never implemented, so it is not regression, they was simply missing.

Patch-set: 2
Label: Verified=0