Update patch set 4
Patch Set 4:

> Given that it's performance improvement and it's internal
 > implementation to the plugin,so why do we want to expose the
 > configuration to enable/disable this feature?

Because it's not performance improvement. Total time spent on both threads (netty and rpc handling thread) will be slighty larger because of creation of new bytebuff and converting it to bytes. This just moves most of the time spent from netty thread to rpc thread, so there may be benefits of disabling this on high-end machines.

Patch-set: 4