Update patch set 6
Patch Set 6: -Code-Review

> > (1 comment)
 > >
 > > Hi Tomas, Although the patch looks good, but have couple of
 > > suggestion.
 > > (1) Fix the headers 2016 --> 2017 for new files
 > > (2) I think we really want to test the serializer/deserializer in
 > a
 > > meaningful way, we should write a single test for each serializer
 > > and deserializer, in which we serialize the data to bytebuf and
 > > then use that bytebuf to deserialize it back to the object and
 > > compare the objects. We should test both positive and negative
 > > cases here.
 > 
 > (1) - Will do, copy/paste error
 > (2) - Thats basically what I am doing, for serializers, I create
 > object to be serialized, serialize it, and then read from byte
 > buffer each field and compare them. And for deserializers, I write
 > field to byte buffer and then compare fields of deserialized
 > object. Also, there is no need for negative cases, because there we
 > are expecting any exception to happen and let openflowjava handle
 > it (it will throw MessageDeserializationFailed with for example
 > NullPointerException etc.) and in case of wrong serialized data,
 > switch will also report error.

If you merge both serializer and deserializer, you will need less code to test both the part and it actually verify that the serialization is correct. You manually checking the serialization doesn't confirm that the serialization is correct. So if skip the manual verification part and let deserialization verify that the serialization is correct, it avoids any manual check + less code and will improve the build time. About the negative test case, yes this is important (not necessary to put in this patch as such), because you should really see that it actually throws the expected exception, like you expect the serialization to write the byte at the correct place. It's not necessary to address both of these comment in here, but if we really want to test serialization in meaningful way, we should merge it. I am removing -1, as these are not really blocker for this patch.

Patch-set: 6
Label: Code-Review=0