Update patch set 14
Patch Set 14:

> It seems odd, from all projects consuming OFP, SFC is the only one
 > making this patch fail the distribution-check.

Yes, it is odd. But SFC have direct dependency on model-flow-base, what depends on openflowjava models, so maybe that is why? Anyway, I tried to rebuild integration/distribution with SFC rebuilt locally with OpenFlowPlugin with this patch and clean .m2, and everything was working fine (without rebuilt SFC, exactly same issue as in logs was happening to me locally when I tried to install odl-sfc-openflow-renderer with Karaf distribution from integration/distribution).

Patch-set: 14