Update patch set 6
Patch Set 6: Code-Review-1

(1 comment)

Hi Tomas, Although the patch looks good, but have couple of suggestion.
(1) Fix the headers 2016 --> 2017 for new files
(2) I think we really want to test the serializer/deserializer in a meaningful way, we should write a single test for each serializer and deserializer, in which we serialize the data to bytebuf and then use that bytebuf to deserialize it back to the object and compare the objects. We should test both positive and negative cases here.

Patch-set: 6