Update patch set 29
Patch Set 29: Code-Review-1

Hi Tomas, Just to give you a context, This issue happens when switch sends flow statistics in multiple multipart response for a single flow statistics request. In the current implementation, plugin delete the old flows from all the tables whenever plugin gets the first multipart response and write the flows of first multipart response to their respective table. The moment second multipart response comes, it again deletes the existing flow from all the tables and writes the flow that is present in the second multipart response. So eventually, you see only flows that is reported through last multipart response is present in the data store. I don't see we are handling that scenario in this patch. Here is the patch that i pushed for the stable/boron for your reference.
https://git.opendaylight.org/gerrit/#/c/50154/9

Patch-set: 29
Label: Code-Review=-1