Update patch set 14
Patch Set 14: Code-Review-1

Hi Tomas,

Can you please add the new test in separate patch so that it can be easier to review. Because last time i reviewed i thought this patch is ready for review, but looks like you added more test and code in it. So please create a new patch for any further testcases.

Patch-set: 14
Label: Code-Review=-1