Update patch set 8
Patch Set 8:

In this modification, the following two types of port-status are not correctly handled. Therefore, I think this patch needs to be updated.

    Problem 1. The port-status which comes before 'initialized' becomes true.

        } else if (initialized) {
            portStatusRegistry.add(portStatus);
        }


    Problem 2. The port-status which comes immediately after handshake completes.

        During handshaking, async messages from a switch are received by OpenflowProtocolListenerInitialImpl. However, all port-status which comes at that time will be discarded.



To handle the problem 1, we are thinking just deleting "if(initialized)" would work fine.
However, we are not sure how to handle the problem 2.
Probably, the following approach would work.

    1. Preserve all port-status which comes into OpenflowProtocolListenerInitialImpl.

    2. When putting OpenflowProtocolListenerFullImpl into OpenflowProtocolListener, copy all port-status in the OpenflowProtocolListenerInitialImpl into DeviceContextImpl.portStatusRegistry.

Patch-set: 8
Label: Verified=0