Update patch set 17
Patch Set 17:

(1 comment)

And yes I agree that is not ideal solution, other solution to this is to add some method to DeviceContext that will update current state every time state is changed in context chain, but I was not sure that I will not miss any state change while doing this (and also no state case, what is important one here) so I just used solution that will guarantee that state will be always correct. But I can rework that part then.

Patch-set: 17