Update patch set 3
Patch Set 3:

Few question Tomas (1) Assuming that during a one stats interval cycle there is not much deletion happened from config data store, so not much change in the flows present in the config data store. In that case, rather then wiping the tables, you are deleting individual flows from each table (even in a transaction chain, but it's a transaction still). So if there are 255 tables with 100 flow each, currently implementation does 1 read( and i assume read is much cheaper than write) + 255 delete and with this patch, you will be doing 255000 deletes, is it right ?
(2) what happen in a scenario where, at P1 stats period it fetches stats and store in the data store, but before it fetches the next  stats, user deletes the flow from the config data store and it's removed from the switch as well. In that case, that flow won't be present in the next set of stats and i assume it's out of flowRegistery as well, so previous stats for that flow will still be in data store ?

I think we need to really check the performance improvement here to make sure it does a real performance improvement. If we don't see any difference, i think we should not merge it as it touches one of the critical code related to stats.

Patch-set: 3