Update patch set 6
Patch Set 6: Code-Review-1

> > Also can you please share your thoughts about following comment
 > > that i made above.
 > >
 > > "Few question Tomas (1) Assuming that during a one stats interval
 > > cycle there is not much deletion happened from config data store,
 > > so not much change in the flows present in the config data store.
 > > In that case, rather then wiping the tables, you are deleting
 > > individual flows from each table (even in a transaction chain,
 > but
 > > it's a transaction still). So if there are 255 tables with 100
 > flow
 > > each, currently implementation does 1 read( and i assume read is
 > > much cheaper than write) + 255 delete and with this patch, you
 > will
 > > be doing 255000 deletes, is it right ?
 > > (2) what happen in a scenario where, at P1 stats period it
 > fetches
 > > stats and store in the data store, but before it fetches the next
 > > stats, user deletes the flow from the config data store and it's
 > > removed from the switch as well. In that case, that flow won't be
 > > present in the next set of stats and i assume it's out of
 > > flowRegistery as well, so previous stats for that flow will still
 > > be in data store ?
 > > I think we need to really check the performance improvement here
 > to
 > > make sure it does a real performance improvement. If we don't see
 > > any difference, i think we should not merge it as it touches one
 > of
 > > the critical code related to stats."
 > 
 > Yes I thought about this, and I am not really sure if this will
 > improve performance, that is why I am testing it right now. One
 > problem with approach we are using currently is that we are doing
 > really big read operation and then big write operation (as we are
 > reading all tables and then basically replacing all of them with
 > same value but without flows) synchronously when we are deleting
 > flows and then proceeding with incoming statistics processing, but
 > this new approach will just prepare this deletion, then prepare
 > writing of flow statistics, but then we will be doing big .submit()
 > to datastore, what might be a problem.
Other issue i see here is that flow-registry is based on the flows present in the config data store and also if there are any alien flow we see from the switch. So there is always a possibility that stats you get from the switch is bit behind then the data you have in the flow registry. So assume a scenario, that you install a flow and that was installed on the switch and then stats were fetched and flow stats were written on the same flow-id. Till now everything is fine, now before the next stats comes, user deletes the flow and it's deleted from the switch as well and will also disappear from the flow registry 
as well. So in this case if we won't wipe the tables, this flow will still be there. Currently we are assuming that flows are permanent flows, but plugin allows timeout flows as well. So if user start using timeout flows, you might end up with multiple deletion and performance might even be bad. Two write flow statistic counters, you will anyways have to write the flows again (we can't merge, because flows can be modified as well). So there are scenario where this patch can gain the performance or it can lose the performance, but irrespective of that, we might see some race conditions as well. So looking at the whole thing, the only performance thing that's worth improving here is avoid doing that big read. I would suggest to use the max-table number that we get with the switch-description. We can make it part of the device context and use that to just wipe the tables. Because i believe this big read we are doing is just for reading the number of tables we have.

Patch-set: 6