Update patch set 20
Patch Set 20:

I think we don't have few test in the CSIT, that we should do to make sure we are not regressing.
(1) Modifying all the related config knobs in openflowplugin.cfg and making sure that those knobs really are picked up.
(2) Disconnecting/Reconnecting switch repeatedly in quick succession and make sure that stats are collected properly. 
(3) Not sure we can test this or not, but i think we should probably test it with a scenario where switch sends multiple response for single multipart request.

Patch-set: 20
Label: Code-Review=-1