Update patch set 11
Patch Set 11:

So after my explanation, is this patch okay for you Anil? It solves all problems with current .cfg file implementation (that is honoring initial values in .yang configuration file, overriding only user-defined settings in .cfg during boot and updating settings on the fly, what is currently on master working, but reloading for few configuration options was missing so I added them, and there was also typo for getting flag for flow-removed-notification, so it was never found in .cfg file even if it was there, so I also fixed that one thing).

Patch-set: 11