Update patch set 3
Patch Set 3:

> > If you are testing with OVS, that won't cover all the scenarios.
 > > With OVS we don't test the multiple multipart response scenario.
 > If
 > > a switch sends the flows stats in more then one multipart reply
 > > message, make sure that you don't cleanup the table while
 > > processing the second multipart response.
 > 
 > It's using transaction chain manager, so this won't be issue,
 > because it is synchronized and everything will be done in order.

Looking at the code, whenever the first multipart reply for stats comes (that contains table 0 stats only), it's deleting all the flows that is present in the flowRegistery, it means it will clear all the tables and write the stats for the table 0 only. Now in the next multipart reply it gets the stats for table 1, then it will again do the same thing and remove all the flows (including the latest stats from table 0) and write table 1 stats. So you end up with stats for only table 1. So i am wondering how transaction chain manager will help here.

Patch-set: 3