Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > > Patch Set 4: Code-Review+1
>  > >
>  > > This works but not sure we can merge in neon anymore...
>  > 
>  > I think we can merge, but it just won't end up in a release unless
>  > there is some
>  > big security fix we need to add that would trigger a security
>  > release.
> 
> I think it is not even building because of the maven central HTTPS issue, and not sure Robert will port the fix to neon.

We will abandon this review and will update the JIRA accordingly. And we can reopen if this change is really necessary in neon. In my opinion, we dont need to keep this review open unless build issues are sorted out and we are not sure if it will be fixed in near future.

Patch-set: 4
Reviewer: Gerrit User 6130 <6130@7fc14799-209e-464c-9743-7a06c2c21a81>