OPNFLWPLUG-1086: Reconciliation framework failure when starting cbench tool for the first time
Issue:

CSIT job using cbench had consistent failure with the error message
starting starting with a "closed connection ... exiting" message.
The karaf logs indicated the switches getting disconnected because of
"reconciliation framework" failure.

Observation and Fix:

On analyzing the logs it was found that the "flownodereconciliation"
service of the reconciliation framework was throwing an uncaught
exception (ConcurrentModificationException) which had resulted in
the reconciliation framework failure.

This has been fixed and logs have been enhanced to debug similar issues.

Change-Id: I3cbcea2da32c342c83846f2041d82e921d7acf08
Signed-off-by: Gobinath <gobinath@ericsson.com>