Update patch set 1
Patch Set 1: -Code-Review

If some cleaning can be done in this patch, why not? If PTL prefers to do it after, that is also fine. I will remove my -1 as this was only to catch attention on the information here, not really to block the patch.

Patch-set: 1
Reviewer: Gerrit User 797 <797@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0