Make AbstractActionSerializer.serialize() final
We want to reduce the number of megamorphic calls in the serialization
path. Rather than overriding the serialize() method in abstract base
class and call into it, make sure the base class calls out to subclasses
to serialize the body.

This reduces the number of serialize() implementations, peeling common
code into the caller. It also makes it clear there are opportunities
to further reduction.

JIRA: OPNFLWPLUG-1102
Change-Id: I0cb421be9f84fd278039a07107254bfac9ddd9a2
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>