Remove DeserializerRegistryInjector
This interface is badly named in that it implies the implementer
is injecting something, whereas it really is an injectee.

Furthermore the injection pattern does not play nicely with SpotBugs,
and further we always know which registry we are going to inject.

Eliminate the interface completely, opting for registry being
supplied to the constructor instead.

Change-Id: Idb5ba098d32ca77d2faf506ccea63ab4f5d7cbee
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>