Migrate openflowplugim-impl tests to uint types
This migrates the entire openflowplugin-impl test suite to use
Uint types.

We borrow EncodeConstants to get at Uint8-specified
protocol numbers, which is probably not what we want, but then
there seems to be variance of opinion whether to use that or
OFConstants. This is fine for this migration, we will address
the OFConstants/EncodeConstants duplicity when we migrate
production code.

JIRA: OPNFLWPLUG-1099
Change-Id: I088597573037b8a662840aff62b0593a058a10b4
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>