Migrate ByteBufUtil.macAddressToString() users
We have users only in tests, perform the straightforward migration.

While we are at it, we also fix expected/actual order in tests, as
well as eliminate call sites which are better served by
Assert.assertArrayEquals().

A few places are asserting the use of upper-case letters, which runs
contrary to what the canonical format is. These are adjusted to use
canonical lower-case.

Change-Id: I4dfc342b005d1cbdec2537eb4cfdac20d6521da2
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>