WIP: Bump upstreams for Silicon
Adopt the following upstream versions:
- odlparent-8.0.0
- yangtools-6.0.0
- mdsal-7.0.0
- controller-3.0.0
- infrautils-1.9.0

Since we are forced to migrate some of the code, perform related
migrations if they are trivial to do.

Quite a few tests are missing proper keys, hence these are added
in the most obvious manner possible. If there is a list/map
warning in that vicinity, also use BindingMap.of() to get rid of
the warning.

Change-Id: Ic4302fba6916dca8385633f1889e8e619fc281e4
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>