Refactor MatchExtensionHelper
This is a rather simple utility class. Since we are cleaning up
deprecation warnings, we might as well completely refactor it.

This results in much more expressive code as well as performance
wins, as we propagate invariants as early as possible.

One lefover is the Optional.ofNullable() thing, which we'll deal
with separately.

Change-Id: Ia92d36626c6c449558a3cd91f4f52b708b0d9645
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>