Drop locking from FlowGroupInfoHistoryImpl
Updates here are in the fast path and we are taking locks for
something that is used ... almost never.

Remove locking and perform time-bound retries to acquire entries
instead, explaining the situation in comments.

If this is not sufficient, a TODO is also added as to what needs
to happen to make the entry acquisition reliable.

Change-Id: I21d05e074bb5033942ef4f7d0bc210a2e9bbf359
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>