Ditch blueprint from srm-impl
We are using a dead plugin to generate half of wiring, with the rest
hand-written. Convert to OSGi DS, noticing a private method, which
really should be part of interface (in keeping with others).

This also fixes a TOCTOU race in removeRecoverableListener(), which
could lead to NPEs.

Since we are now SCR-enabled, we can ditch the use of ODL blueprint
extensions in shell, making everyone a bit happier.

Change-Id: Ia9d4d0fc3f0bc430d3a106f53975ab7d6e1241d0
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>