WIP: Decompose RPC implementation classes
RpcContextImpl registers a multitude of RPC services when it becomes
active. We really want to make that registration atomic -- which is
something we can do with
RpcProviderService.registerRpcImplementations().

This patch splits up our implementation classes to individual RPCs and
then uses a single call to register them in one go.

For bonus points we get rid of the horrible java.lang.IllegalAccessError
throws and correctly report RPC unavailability (because we do not
register those RPCs).

We also ditch OpendaylightDirectStatisticsServiceProvider and its
associated *LayerDirectStatisticsProviderInitializer, as they can be
cleanly bound without per-request lookups.

// FIXME: OpendaylightQueueStatisticsService
// FIXME: OpendaylightPortStatisticsService

// FIXME: fixup test references

JIRA: OPNFLWPLUG-1125
Change-Id: I80ade2f9f51fc765c0d685f544e5b102061d033c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>