Expose DeviceFlowRegistry.createKey()
We are lugging around protocol version, which is already known to
DeviceFlowRegistry implementation only to circle back to a utility
method, which then performs lookup and static dispatch.

Short circuit this dance by exposing createKey(), which allows
DeviceFlowRegistryImpl to cleanly integrate with rest of the world.

All sorts of aspects are addressed as well:
- use a record for the canonical key implementation
- use well-known constants for FlowRegistryKeyFactory and simple if/else
  dispatch
- make MatchNormalizationUtil a class capturing its associated version

Future improvements are marked with FIXMEs, as these will further
reduce reliance on constant protocol version passing.

Change-Id: I70d22f0db65b97df4b63c2bc1bf4cff457a11785
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>