Improve DeviceMastershipManager locking
Use a concurrent implementation of Set internally, which allows us to
perform add/remove concurrently -- and we can use the return value to
know when a device was added/removed.

While we are here, ditch the double lookup in setNodeOperationalStatus()
and issue a get() with a null check.

Furthermore convert lockObj to an explicit lock, as that is a more
friendly thing to do.

Change-Id: Ib5589db4cdb35259f6b400091ec2817c6b0751cc
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>