Refactor southbound-cli
We have a rather weird wiring here: there are commands, but there is
also a YANG model, which exposes a single, multiplexing, RPC.

Ditch the YANG RPCs and the active component, shifting the logic into
the CLI command only.

This removes quite a bit of complexity in terms of wiring and code
organization.

// FIXME: argh, this needs more work

JIRA: OPNFLWPLUG-1112
JIRA: OPNFLWPLUG-1125
Change-Id: I17520ca03a153ff64af264a88b41bccc4e108857
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>