Refactor ReconciliationServiceImpl
This service implements a single RPC, Reconcile, and is injected into
its sole user.

Defile a separate interface, ReconcileService, which defines the two
possibilities the RPC can handle.

Rename ReconciliationServiceImpl to DefaultReconcileService and let it
register the RPC itself. While we are at it, convert it to OSGi DS.

JIRA: OPNFLWPLUG-1112
JIRA: OPNFLWPLUG-1125
Change-Id: Ifc9d3707191484b492ebf5a190f65d5b673a5bdc
Signed-off-by: lubos-cicut <lubos.cicut@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>