Add OSGiConfiguration
We are using properties files for primary configuration, coupled to
Config Admin in OSGi environment.

This patch adds OSGiConfiguration, which is an ObjectClassDefinition of
everything we support under
configurationPid="org.opendaylight.openflowplugin". It lays down the
foundation for transitioning to properly type-safe configuration.

It also relocates OFConstants.CONFIG_FILE_ID, as it really is an
implementation-specific detail.

JIRA: OPNFLWPLUG-1132
Change-Id: I695d98dca8032ae91fb3a6501e2ff43ab007fb44
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>