Wire OpenFlowPluginProviderImpl via OSGi DS
OpenFlowPluginProviderImpl has interactions with injected
SwitchConnectionProviders which get interacted with during shutdown.

This is problematic if Blueprint fails to notice it is shutting down
and there is no point in listening for dependencies showing up in the
proxy interfaces.

While the SwitchConnectionProvider lifecycle needs a deep look, take
advantage of the fact that OPNFLWPLUG-1129 laid the SCR ground work
and move OpenFlowPluginProviderImpl to SCR as well.

This takes Blueprint out of the picture, with SCR guaranteeing we are
looking at a live SwitchConnectionProvider when shutting down.

Unfortunately we still have one more service in Blueprint, hence we need
to manually specify Provides manifest header. That part will be
addressed in a follow-up patch.

JIRA: OPNFLWPLUG-1132
Change-Id: I5329c3209342f52caf7b4403f5c0bcff65c9f669
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit 271bb499582603561140828ae2b044dba2f12a2a)