Refactor SwitchConnectionProvider a bit
This interface specifies AutoCloseable, which is an implementation leak
from the days of Config Subsystem. Eliminate this specification and make
sure the close() method performs deactivation.

This is the step in the direction of having well-defined lifecycle,
where the users cannot interfere with what the provider is doing.

Change-Id: I1ccbb6987f21b3871d80a9da694211e3c3d19204
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit b753a6614a2036e78f76d7c17fde521a58f887da)