Clean up NodeConnectorInventoryEventTranslator
This entire class is quite confused, using two separate
DataTreeChangeListeners, going through quite the hoops to try to discern
proper activations -- and failing anyway.

All of this logic is an evolution of DataChangeListener from the first
cut at binding-api, which were not properly cleaned up to do
DataTreeChangeListener logic.

This patch cleans all of that up by using a single listener and just
checking the after-image to see if the NodeConnector is up. We also keep
proper track of what we have reported, so that we up/down/up and similar
transitions are completely tracked.

Change-Id: Ia91587c324418d1a7be9cc14b89a2d540f694f90
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>