Remove NetUtils methods
Just like any other FooUtils dumping ground of a class, NetUtils
contains an absolutely bonkers number of methods that are not used from
anywhere, which is to say they do not even have tests.

Mass remove everything that is not referenced -- which leaves us with
exactly two methods, used from exactly one class in the same package.

Change-Id: I2890d5e3638d92543ab8eb849872d34407c3e5b5
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>